refactor: Extract locale management and improve pint_to_sympy
This refactor improves code organization and performance:

**Phase 1: Locale Management Extraction**
- Created src/keecas/localization/pint_locale.py
- Moved all locale functions from pint_sympy.py
- Cleaner separation of concerns

**Phase 2: Core Refactor**
- Added SymPyUnitCache class for unit caching
- Extracted _is_unit_prefixed() helper function
- Refactored pint_to_sympy with clearer structure
- Removed obsolete commented code

**Phase 3: Cleanup**
- Updated test imports
- All 109 tests passing
- No breaking changes to public API

**Impact:**
- 65% LOC reduction in pint_sympy.py (462 → 160 lines)
- Better performance via unit caching
- Improved maintainability and readability
